% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03_eval_results.R
\name{eval_results}
\alias{eval_results}
\title{Evaluate model performance for each sub-period}
\usage{
eval_results(results_cal, results_test, events, Qprob_thr_P1_P2)
}
\arguments{
\item{results_cal}{A tibble containing the model results in calibration mode.}

\item{results_test}{A tibble containing the model results in evaluation mode.}

\item{events}{A tibble containing information about the flow events: start, end dates, and event IDs}

\item{Qprob_thr_P1_P2}{Streamflow threshold to evaluate contingency table}
}
\value{
A tibble containing the calculated performance metrics for the specified period, including:
\itemize{
\item \verb{NSE_KGE bounded}: NSE and KGE metrics for overall performance and events.
\item \code{cont_table}: Contingency table metrics - POD, FAR, CSI, for each model variable, Qprev, Qcorr, Qtan
}
}
\description{
This function evaluates the performance of GR5H_RI model and the post-processing model results over different
based on various performance metrics, such as Nash-Sutcliffe Efficiency coefficient and a contingency table for flood events.
}
