% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate_results.R
\name{evaluate_results}
\alias{evaluate_results}
\title{Evaluate model performance (NSE, KGE, C2MP, POD, FAR, CSI)}
\usage{
evaluate_results(results_cal, results_test, events, Qthr, Hprev)
}
\arguments{
\item{results_cal}{A tibble containing the model results in calibration mode.}

\item{results_test}{A tibble containing the model results in evaluation mode.}

\item{events}{A tibble containing information about the flow events: start, end dates, and event IDs}

\item{Qthr}{Streamflow threshold to evaluate contingency table}

\item{Hprev}{Forecast horizon ("H3", "H6", "H12", "H24").}
}
\value{
A tibble containing the calculated performance metrics for the specified period, including:
\itemize{
\item \verb{NSE_KGE bounded}: NSE and KGE metrics for overall performance and flood events.
\item \code{C2MP}: C2MP metrics for overall performance and flood events.
\item \code{cont_table}: Contingency table metrics - POD, FAR, CSI, for each model variable, Qprev, Qcorr, Qtan
}
}
\description{
Evaluate model performance (NSE, KGE, C2MP, POD, FAR, CSI)
}
