% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02_post_processing.R
\name{post_processing}
\alias{post_processing}
\title{Post-processing Model Application}
\usage{
post_processing(input_data_cal, input_data_eval, model_name)
}
\arguments{
\item{input_data_cal}{A tibble containing the pre-processed GR5H_RI forecast data for the calibration period.}

\item{input_data_eval}{A tibble containing the pre-processed GR5H_RI forecast data for the evaluation period.}

\item{model_name}{A string indicating the type of post-processing model to apply ("mlp", "random_forest" or "xgboost").}
}
\value{
A list containing:
\itemize{
\item \code{cal}: The model's predictions for the calibration period.
\item \code{test}: The model's predictions for the evaluation period.
\item \code{model}: Fitted model and calibrated hyperparameters.
}
}
\description{
This function applies the post-processing model (MLP, Random Forest, or XGBoost) to the pre-processed GR5H_RI forecast data.
}
