% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01_prepare_input_data.R
\name{prepare_input_data}
\alias{prepare_input_data}
\title{Prepare Input Data for Post-Processing Model}
\usage{
prepare_input_data(path_prev, time_steps)
}
\arguments{
\item{path_prev}{Path to the GR5H_RI forecast data.}

\item{time_steps}{Time window, in hours, used to compute past streamflow prediction errors for the post-processing model.}
}
\value{
A list containing:
\itemize{
\item \code{prev}: A prepared dataset, containing processed data for use in the post-processing model, including time series
of observed and forecasted streamflow, accumulated precipitation and predicted streaflow errors.
\item \code{Qprob_thr}: The streamflow threshold that will be used for flood event identification in the contingency table.
}
}
\description{
This function prepares input data from GR5H_RI forecasts and observed data for use in a post-processing model.
}
\details{
The following variables must be defined before using this function:
\itemize{
\item Qobs_ts: Observed streamflow time series tibble with columns "Date" (dttm) and "Qobs" (dbl).
\item Pmm: Precipitation time series tibble with columns "Date" (ddtm) and "Pmm" (dbl).
\item time_steps: time window, in hours, used to compute past streamflow prediction errors for the post-processing model.
\item Hcal: calibration horizon ("H3", "H6", "H12", "H24")
\item Hprev: forecast horizon ("H3", "H6", "H12", "H24")
}
}
